Reestruturacao de codigo
