+++ salic-api/resources/projeto/Segmento.py
@@ -1,24 +1,23 @@
 from ..result_format import get_formated, to_json
 from flask_restful import Api
-from database.QueryHandler import QueryHandler
 from flask import Response
 from ..ResourceBase import *
+from models import SegmentoModelObject
 
 class Segmento(ResourceBase):
-    
+
     def __init__(self):
         super (Segmento,self).__init__()
-        
-    
+
+
     def get(self):
-        
+
         try:
-            result = self.query_handler.get_segmento()
+            result = SegmentoModelObject().all()
         except:
             result =  { 'message' : 'internal error',
                         'message_code' : 13
                         }
             return self.render(result, status_code = 503)
-        
+
         return self.render(result)
-
